pyproject.toml
[tool.poetry]
name = "bpm-sorter"
version = "0.1.0"
description = "Sort Spotify playlists by their BPM"
authors = ["Tim Jentzsch <bpm_sorter.projects@timjen.net>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "D", "I"]
lint.ignore = [
  # multi-line-summary-second-line
  # Collides with other lint rule and is non-standard
  "D213",
  # one-blank-line-before-class
  # Collides with other lint rule and is non-standard
  "D203",
  # undocumented-public-init
  # Often just repeats the class name and doesn't really add new information
  "D107",
]

[tool.poetry.dependencies]
python = "^3.10"
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
mypy = "^1.11.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
